add groupcache tests

+++ tools/test.sh
@@ -52,15 +52,17 @@ S3=$(tempfile); rm $S3
 L1=$(tempfile)
 L2=$(tempfile)
 L3=$(tempfile)
+L4=$(tempfile)
 E1=$(tempfile)
 E2=$(tempfile)
 E3=$(tempfile)
+E4=$(tempfile)
 
 PORT=$((9999 + $$))
 
 knxd -n K1 -t 0xfffc -f 9 -e 3.2.1 -E 4.3.2:5 -c -u$S1 -u$S2 -DTR --Server=:$PORT -bdummy: &
 KNX1=$!
-trap 'rm -f $L1 $L2 $E1 $E2 $EF; kill $KNX1' 0 1 2
+trap 'echo T1; rm -f $L1 $L2 $E1 $E2 $EF; kill $KNX1; wait' 0 1 2
 
 sleep 1
 #echo knxd -t 0xfffc -f 9 -e 3.2.2 -E 4.5.6:10 -i$((10002 + $$)) -u$S3 -b ipt:localhost:$PORT:$((10000 + $$)) 
@@ -68,7 +70,7 @@ sleep 1
 knxd -n K2 -t 0xffff -f 9 -e 3.2.2 -E 4.5.6:5 -i$((10002 + $$)) -u$S3 -b ipt:localhost:$PORT:$((10000 + $$)) &
 KNX2=$!
 #read RETURN
-trap 'rm -f $L1 $L2 $E1 $E2 $EF; kill $KNX1 $KNX2' 0 1 2
+trap 'echo T2; rm -f $L1 $L2 $E1 $E2 $EF; kill $KNX1 $KNX2; wait' 0 1 2
 sleep 1
 
 knxtool grouplisten local:$S2 1/2/3 >$L1 2>$E1 &
@@ -77,28 +79,34 @@ knxtool vbusmonitor1 local:$S2 >$L2 2>$E2 &
 PL2=$!
 knxtool vbusmonitor1 local:$S3 >$L3 2>$E3 &
 PL3=$!
+knxtool groupcacheread local:$S3 1/2/3 2 >$L4 2>$E4 &
+PL4=$!
 # will die by itself when the server terminates
 
 # test that addresses get recycled
 sleep 1
-knxtool groupswrite local:$S1 1/2/3 4
+if ! knxtool groupswrite local:$S1 1/2/3 4 ; then echo X1; exit 1; fi
 sleep 1
-knxtool groupswrite local:$S1 1/2/3 4
+if ! knxtool groupswrite local:$S1 1/2/3 5 ; then echo X2; exit 1; fi
 sleep 1
-knxtool groupswrite local:$S1 1/2/3 4
+if ! knxtool groupswrite local:$S1 1/2/3 6 ; then echo X3; exit 1; fi
 sleep 1
-knxtool groupswrite local:$S1 1/2/3 4
+if ! knxtool groupswrite local:$S1 1/2/3 7 ; then echo X4; exit 1; fi
 sleep 1
-knxtool groupswrite local:$S1 1/2/3 4
+if ! knxtool groupswrite local:$S1 1/2/3 8 ; then echo X5; exit 1; fi
 sleep 1
-knxtool groupwrite local:$S2 1/2/3 4 5 6
+if ! knxtool groupwrite local:$S2 1/2/3 4 5 6 ; then echo X6; exit 1; fi
+sleep 1
+kill $PL4 || true
+if ! knxtool groupcacheread local:$S1 1/2/3 >>$L4 2>>$E4 ; then echo X7; exit 1;
+fi
+if ! knxtool groupcachelastupdates local:$S1 3 1 >>$L4 2>>$E4 ; then echo X7; exit 1; fi
 
-sleep 1
 #read RETURN
 kill $KNX1 $KNX2
 sleep 1
 kill $PL1 $PL2 $PL3 || true
-trap 'rm -f $L1 $L2 $L3 $E1 $E2 $E3 $EF' 0 1 2
+trap 'echo T3; rm -f $L1 $L2 $L3 $L4 $E1 $E2 $E3 $E4 $EF' 0 1 2
 sleep 1
 #ls -l $L1 $L2 $E1 $E2
 #cat $L1 $L2 $E1 $E2
@@ -110,6 +118,7 @@ E=""
 diff -u "$(dirname "$0")"/logs/listen $L1 || E=1$E
 diff -u "$(dirname "$0")"/logs/monitor $L2 || E=2$E
 diff -u "$(dirname "$0")"/logs/monitor2 $L3 || E=3$E
+diff -u "$(dirname "$0")"/logs/cache $L4 || E=4$E
 test -z "$E"
 
 set +ex
